MFC r346701: loader: fdt: Add fdt_is_setup function
When efi_autoload is called it will call fdt_setup_fdtp which setup the
dtb and overlays. If a user already loaded at dtb or overlays or just
printed the efi provided dtb, this will re-setup everything and also
re-applying the overlays.
Test that everything is setup before doing it again.


git-svn-id: https://svn.freebsd.org/base/stable/12@353980 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f