MFC r349471, r351166: Tweak EFI_STAGING_SIZE
r349471:
Increase EFI_STAGING_SIZE to 100MB on x64

To avoid failures when the large 18MB nvidia.ko module is being loaded,
increase EFI_STAGING_SIZE from 64MB to 100MB on x64 systems.
Leave the other platforms at 64MB.

r351166:
Reduce size of EFI_STAGING_SIZE to 32 on arm

Reduce the size of the EFI_STAGING area we allocate on arm to 32. On arm SBC
such as the NanoPi-NEOLTS the staging area allocation will fail on the 256MB
model with a staging size of 64.


git-svn-id: https://svn.freebsd.org/base/stable/12@354009 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f