MFC 353023: Fix the EMBEDFS_FORMAT helper variable for riscv64.
It was defined with the wrong MACHINE_ARCH previously.  This permits
using an MFS image without defining MD_ROOT_SIZE which has various
benefits (one being that the build is able to treat the MFS image as
a dependency and properly re-link the kernel with the new image when
building with NO_CLEAN).


git-svn-id: https://svn.freebsd.org/base/stable/11@354109 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f