MFC r353334:
RISC-V: Fix an alignment warning in libthr

Compiling with clang gives a loss-of-alignment error due the cast to
uint8_t *. Since the TLS is always tcb aligned and TP_OFFSET is defined
as sizeof(struct tcb) we can guarantee there is no misalignment. Silence
the error by moving the offset into the inline assembly.

Reviewed by:	br
Differential Revision:	https://reviews.freebsd.org/D21926


git-svn-id: https://svn.freebsd.org/base/stable/12@354263 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f