[ig4] Add support for polled mode
Currently ig4 internally depends on it's own interrupts and uses mtx_sleep()
to wait for them. That means it can not be used from any context where
sleeping is disallowed e.g. on cold boot, from DDB/KDB, from other device
driver's interrupt handlers and so on.

This change replaces sleeps with busy loops in cold boot and DDB cases.


git-svn-id: https://svn.freebsd.org/base/head@354301 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f