[ig4] Do not wait until interrupts are enabled at attach stage
as the driver is fully functional on a cold boot through utilization of
polled mode.

As a side effect, ig4 children probe and attach methods can be called
earlier in the boot sequence, so now it is up to the child drivers
to wait for a kernel initialization completion if it is required.


git-svn-id: https://svn.freebsd.org/base/head@354303 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f