Create a new sysctl subtree, machdep.mitigations.  Its purpose is to organize knobs and indicators for code that mitigates functional and security issues in the architecture/platform.  Controls for regular operational policy should still go into places security, hw, kern, etc.
The machdep root node is inherently architecture dependent, but mitigations
tend to be architecture dependent as well.  Some cases like Spectre do cross
architectural boundaries, but the mitigation code for them tends to be
architecture dependent anyways, and multiple architectures won't be active
in the same image of the kernel.

Many mitigation knobs already exist in the system, and they will be moved
with compat naming in the future.  Going forward, mitigations should collect
in machdep.mitigations.

Reviewed by:	imp, brooks, rwatson, emaste, jhb
Sponsored by:	Intel


git-svn-id: https://svn.freebsd.org/base/head@354756 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f