MFC r351812: mpsutil slot set status
It allows to set the status of an enclosure slot.  Practically, this
means controlling whatever slot status LEDs the enclosure provides.  At
present, the new command does not have sanity checks or any
conveniences.  That means that it is possible to issue the command for
an invalid slot and an enclosure.  But the worst I have seen happening
is either the command failing or simply being ignored.  Also, at the
moment, the status has to be specified as a numeric bit mask.  The bit
definitions can be found in sys/dev/mps/mpi/mpi2_init.h, they are
prefixed with MPI2_SEP_REQ_SLOTSTATUS_.  The only way to address a slot
is by the enclosure handle and the slot number.  Both are readily
available from mpsutil show commands.

The new command is useful alternative to 'sas2ircu locate' command.
First, sas2ircu is a proprietary blob.  Second, it supports setting only
locate / identify status bit.


git-svn-id: https://svn.freebsd.org/base/stable/12@354884 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f